avcodec/hq_hqadsp: Fix runtime error: signed integer overflow: 80359 * 30274 cannot be represented in type 'int'
Fixes: 1507/clusterfuzz-testcase-minimized-4955228300378112

Found-by: continuous fuzzing process https://github.com/google/oss-fuzz/tree/master/targets/ffmpeg

+++ libavcodec/hq_hqadsp.c
@@ -30,7 +30,7 @@
 #define FIX_1_414 23170
 #define FIX_2_613 21407 // divided by two to fit the range
 
-#define IDCTMUL(a, b) ((a) * (b) >> 16)
+#define IDCTMUL(a, b) ((int)((a) * (unsigned)(b)) >> 16)
 
 static inline void idct_row(int16_t *blk)
 {
